+++ MaterialDesignExtensions/Controls/FileSystemControl.cs
@@ -33,6 +33,7 @@ namespace MaterialDesignExtensions.Controls
         protected const string DrawerHostName = "drawerHost";
         protected const string PathPartsScrollViewerName = "pathPartsScrollViewer";
         protected const string PathPartsItemsControlName = "pathPartsItemsControl";
+        protected const string CurrentDirectoryTextBoxName = "currentDirectoryTextBox";
         protected const string FileSystemEntryItemsControlName = "fileSystemEntryItemsControl";
         protected const string EmptyDirectoryTextBlockName = "emptyDirectoryTextBlock";
 
@@ -203,6 +204,31 @@ namespace MaterialDesignExtensions.Controls
         }
 
         /// <summary>
+        /// Show the current directory path with a short cut button for each part in the path, instead of a text box.
+        /// </summary>
+        public static readonly DependencyProperty PathPartsAsButtonsProperty = DependencyProperty.Register(
+            nameof(PathPartsAsButtons),
+            typeof(bool),
+            typeof(FileSystemControl),
+            new PropertyMetadata(true));
+
+        /// <summary>
+        /// Show the current directory path with a short cut button for each part in the path, instead of a text box.
+        /// </summary>
+        public bool PathPartsAsButtons
+        {
+            get
+            {
+                return (bool)GetValue(PathPartsAsButtonsProperty);
+            }
+
+            set
+            {
+                SetValue(PathPartsAsButtonsProperty, value);
+            }
+        }
+
+        /// <summary>
         /// Shows or hides hidden directories and files.
         /// </summary>
         public static readonly DependencyProperty ShowHiddenFilesAndDirectoriesProperty = DependencyProperty.Register(
@@ -277,10 +303,36 @@ namespace MaterialDesignExtensions.Controls
             }
         }
 
+        /// <summary>
+        /// Enable switching between a text box and the buttons for each sub directory of the current directory's path.
+        /// </summary>
+        public static readonly DependencyProperty SwitchPathPartsAsButtonsEnabledProperty = DependencyProperty.Register(
+            nameof(SwitchPathPartsAsButtonsEnabled),
+            typeof(bool),
+            typeof(FileSystemControl),
+            new PropertyMetadata(false));
+
+        /// <summary>
+        /// Enable switching between a text box and the buttons for each sub directory of the current directory's path.
+        /// </summary>
+        public bool SwitchPathPartsAsButtonsEnabled
+        {
+            get
+            {
+                return (bool)GetValue(SwitchPathPartsAsButtonsEnabledProperty);
+            }
+
+            set
+            {
+                SetValue(SwitchPathPartsAsButtonsEnabledProperty, value);
+            }
+        }
+
         protected FileSystemController m_controller;
 
         protected ScrollViewer m_pathPartsScrollViewer;
         protected ItemsControl m_pathPartsItemsControl;
+        protected TextBox m_currentDirectoryTextBox;
         // use an ItemsControl instead of a ListBox, because the ListBox raises several selection changed events without an explicit user input
         // another advantage: non selectable items such as headers can be added
         protected ItemsControl m_fileSystemEntryItemsControl;
@@ -303,6 +355,7 @@ namespace MaterialDesignExtensions.Controls
             m_controller.SelectDirectory(CurrentDirectory);
 
             CommandBindings.Add(new CommandBinding(FileSystemControlCommands.OpenSpecialDirectoriesDrawerCommand, OpenSpecialDirectoriesDrawerCommandHandler));
+            CommandBindings.Add(new CommandBinding(FileSystemControlCommands.SwitchPathPartsAsButtonsCommand, SwitchPathPartsAsButtonsHandler));
             CommandBindings.Add(new CommandBinding(FileSystemControlCommands.SelectDirectoryItemCommand, SelectDirectoryItemCommandHandler));
             CommandBindings.Add(new CommandBinding(FileSystemControlCommands.SelectFileSystemEntryCommand, SelectFileSystemEntryCommandHandler));
             CommandBindings.Add(new CommandBinding(FileSystemControlCommands.ShowInfoCommand, ShowInfoCommandHandler));
@@ -315,6 +368,7 @@ namespace MaterialDesignExtensions.Controls
 
             m_pathPartsScrollViewer = null;
             m_pathPartsItemsControl = null;
+            m_currentDirectoryTextBox = null;
             m_fileSystemEntryItemsScrollViewer = null;
             m_fileSystemEntryItemsControl = null;
 
@@ -331,6 +385,9 @@ namespace MaterialDesignExtensions.Controls
             m_pathPartsItemsControl = Template.FindName(PathPartsItemsControlName, this) as ItemsControl;
             m_pathPartsItemsControl.ItemsSource = m_controller.CurrentDirectoryPathParts;
 
+            m_currentDirectoryTextBox = Template.FindName(CurrentDirectoryTextBoxName, this) as TextBox;
+            m_currentDirectoryTextBox.Text = CurrentDirectory;
+
             m_fileSystemEntryItemsControl = Template.FindName(FileSystemEntryItemsControlName, this) as ItemsControl;
             m_fileSystemEntryItemsControl.ItemsSource = GetFileSystemEntryItems();
 
@@ -342,6 +399,7 @@ namespace MaterialDesignExtensions.Controls
         protected virtual void LoadedHandler(object sender, RoutedEventArgs args)
         {
             m_controller.PropertyChanged += ControllerPropertyChangedHandler;
+            m_currentDirectoryTextBox.KeyDown += CurrentDirectoryTextBoxKeyDownHandler;
 
             // not sure if the control should get keyboard focus on loading
             //Keyboard.Focus(this);
@@ -350,12 +408,35 @@ namespace MaterialDesignExtensions.Controls
         protected virtual void UnloadedHandler(object sender, RoutedEventArgs args)
         {
             m_controller.PropertyChanged -= ControllerPropertyChangedHandler;
+            m_currentDirectoryTextBox.KeyDown -= CurrentDirectoryTextBoxKeyDownHandler;
         }
 
         protected void OpenSpecialDirectoriesDrawerCommandHandler(object sender, ExecutedRoutedEventArgs args)
         {
             DrawerHost drawerHost = ((DrawerHost)Template.FindName(DrawerHostName, this));
             drawerHost.IsTopDrawerOpen = true;
+        }
+
+        protected void SwitchPathPartsAsButtonsHandler(object sender, ExecutedRoutedEventArgs args)
+        {
+            PathPartsAsButtons = !PathPartsAsButtons;
+        }
+
+        private void CurrentDirectoryTextBoxKeyDownHandler(object sender, KeyEventArgs args)
+        {
+            if (args.Key == Key.Enter)
+            {
+                string directory = m_currentDirectoryTextBox.Text
+                    .Replace("\n", string.Empty)
+                    .Replace("\r", string.Empty)
+                    .Replace("\r", string.Empty)
+                    .Trim();
+
+                if (!string.IsNullOrWhiteSpace(directory))
+                {
+                    CurrentDirectory = directory;
+                }
+            }
         }
 
         protected void SelectDirectoryItemCommandHandler(object sender, ExecutedRoutedEventArgs args)
@@ -460,6 +541,7 @@ namespace MaterialDesignExtensions.Controls
             try
             {
                 m_controller.SelectDirectory(newCurrentDirectory);
+                m_currentDirectoryTextBox.Text = newCurrentDirectory;
             }
             catch (PathTooLongException)
             {